modify related data removal, add quick client or car adding and order creation after that

+++ app/Http/Controllers/Admin/Category/DestroyController.php
@@ -9,7 +9,13 @@ class DestroyController extends Controller
 {
     public function __invoke(Category $category)
     {
-        $category->delete();
+        if($category->tasks->isEmpty()) {
+            $category->delete();
+        } else {
+            session()->flash('error', 'В данной категории имеются работы. Удаление категории невозможно!');
+            return view('admin.categories.show', compact('category'));
+        }
+
         return redirect()->route('admin.categories.index');
     }
 }
