+++ app/Http/Controllers/Admin/Task/DestroyController.php
@@ -9,7 +9,12 @@ class DestroyController extends Controller
 {
     public function __invoke(Task $task)
     {
-        $task->delete();
+        if($task->orders->isEmpty()) {
+            $task->delete();
+        } else {
+            session()->flash('error', 'Имеются заказы с данной работой. Удаление работы невозможно!');
+            return view('admin.tasks.show', compact('task'));
+        }
         return redirect()->route('admin.tasks.index');
     }
 }
