+++ app/Http/Controllers/Admin/User/Car/DestroyController.php
@@ -11,7 +11,13 @@ class DestroyController extends Controller
 {
     public function __invoke(User $user, Car $car)
     {
-        $car->delete();
+        if($car->orders->isEmpty()) {
+            $car->delete();
+        } else {
+            session()->flash('error', 'По данному авто имеются заказы. Удаление автомобиля невозможно!');
+            return view('admin.users.cars.show', compact('user', 'car'));
+        }
+
         return redirect()->route('admin.users.cars.index', $user->id);
     }
 }
